ref(xmpp): adjust last success time tracker for bosh and websocket (#1053)
